IGNITE-22624 Fix compilation and move introduce new TableAwareReplicaRequestPreProcessor
